requirements.txt
pdfplumber
chromadb
sentence-transformers
fastapi
uvicorn
requests
python-dotenv
streamlit
